requirements.txt
feedparser==6.0.10
openai>=1.30.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
flask==3.0.3
